requirements.txt
python-dotenv==1.1.1
google-generativeai==0.8.5
playwright==1.54.0
requests==2.32.4
beautifulsoup4==4.13.4
aiohttp==3.9.3
pytest==8.0.2
loguru==0.7.2
pyyaml==6.0.1
requests==2.31.0
undetected-chromedriver==3.5.5
selenium==4.18.1
pillow==10.2.0
opencv-python==4.9.0.80 